refactor: optimize ajv load order
