+++ packages/core-browser/src/layout/index.ts
@@ -1,3 +1,5 @@
 export * from './layout';
+export * from './layout-hooks';
+export * from './layout.interface';
 export * from './accordion/view-context-key.registry';
 export * from './accordion/tab-bar-toolbar';
