+++ packages/debug/__tests__/browser/view/breakpoint/debug-breakpoints.service.test.ts
@@ -8,9 +8,8 @@ import { createBrowserInjector } from '@opensumi/ide-dev-tool/src/injector-helpe
 import { MockInjector } from '@opensumi/ide-dev-tool/src/mock-injector';
 import { IEditorDocumentModelService } from '@opensumi/ide-editor/lib/browser';
 import { MockFileServiceClient } from '@opensumi/ide-file-service/lib/common/mocks';
-import { IWorkspaceService } from '@opensumi/ide-workspace';
+import { IWorkspaceService, IWorkspaceStorageService } from '@opensumi/ide-workspace';
 import { WorkspaceEditDidRenameFileEvent, WorkspaceEditDidDeleteFileEvent } from '@opensumi/ide-workspace-edit';
-import { IWorkspaceStorageService } from '@opensumi/ide-workspace/lib/common/workspace.definition';
 
 describe('Debug Breakpoints Service', () => {
   const mockInjector = createBrowserInjector(
