+++ packages/preferences/__tests__/common/workspace.definition.test.ts
@@ -1,5 +1,6 @@
 import { URI, FileStat } from '@opensumi/ide-core-browser';
-import { DEFAULT_WORKSPACE_SUFFIX_NAME, WorkspaceData } from '@opensumi/ide-workspace';
+import { DEFAULT_WORKSPACE_SUFFIX_NAME } from '@opensumi/ide-workspace';
+import { WorkspaceData } from '@opensumi/ide-workspace/lib/browser/workspace-data';
 
 describe('WorkspaceData methods', () => {
   it('is method should be work', () => {
