+++ packages/preferences/src/browser/workspace-file-preference-provider.ts
@@ -1,7 +1,8 @@
 import { Autowired, Injectable } from '@opensumi/di';
 import { URI } from '@opensumi/ide-core-browser';
 import { PreferenceScope } from '@opensumi/ide-core-browser/lib/preferences';
-import { IWorkspaceService, WorkspaceData } from '@opensumi/ide-workspace';
+import { IWorkspaceService } from '@opensumi/ide-workspace';
+import { WorkspaceData } from '@opensumi/ide-workspace/lib/browser/workspace-data';
 
 import { AbstractResourcePreferenceProvider } from './abstract-resource-preference-provider';
