+++ packages/terminal-next/src/browser/links/validated-local-link-provider.ts
@@ -10,7 +10,7 @@ import { Injectable, Autowired, INJECTOR_TOKEN, Injector } from '@opensumi/di';
 import { AppConfig } from '@opensumi/ide-core-browser/lib/react-providers/config-provider';
 import { IWindowService } from '@opensumi/ide-core-browser/lib/window';
 import { CommandService, IDisposable, URI, OperatingSystem } from '@opensumi/ide-core-common';
-import { IWorkspaceService } from '@opensumi/ide-workspace/lib/common/workspace.definition';
+import { IWorkspaceService } from '@opensumi/ide-workspace/lib/common/workspace.interface';
 
 import type { TerminalClient } from '../terminal.client';
