+++ packages/workspace/src/common/constants.ts
@@ -1,0 +1,4 @@
+export const DEFAULT_WORKSPACE_SUFFIX_NAME = 'sumi-workspace';
+export const WORKSPACE_USER_STORAGE_FOLDER_NAME = '.sumi';
+export const WORKSPACE_RECENT_DATA_FILE = 'recentdata.json';
+export const UNTITLED_WORKSPACE = 'Untitled';
