+++ packages/workspace/src/common/index.ts
@@ -1,1 +1,2 @@
-export * from './workspace.definition';
+export * from './constants';
+export * from './workspace.interface';
